requirements.txt
fastapi
uvicorn
python-docx
pymupdf
requests
beautifulsoup4
pytest
streamlit
agno
ollama
